Integrate sbperfmonitor as implemented in datacollector
